Move Alexa send* functions from smarthome/client.ts to smarthome/index.ts
client.ts now only exposes OAuth token management and makeEventsRequest.
The three outbound event helpers (sendAddOrUpdateReport, sendDeleteReport,
sendSimpleEventSource) live in smarthome/index.ts alongside the handlers
that use them, making client.ts an internal detail of the smarthome/ folder.
